Move healthchecks to a separate file
